Create factory function and gameboard functions
I created a factory function to make the players and set their marks to
either 'O' or 'X'. I also added some functions to the gameboard module
such as PlaceX and PlaceO which will place the corresponding mark on an
empty gameboard tile. I also added a restart button which clears the
gameboard and a startNewGame() function. Will definitely need to refactor
this code at some point but just working on getting the functionality
down at the moment.