Define placeMark function and switch player logic
I changed the original placeX and placeO functions with a general
placeMark function which takes in the tile to change and the mark to
change it to. I also created a switchPlayer function which swaps the
currentPlayer variable and allows the alternating placement of Xs and Os.
This function is contained within the placeMark function and is not
returned at the end of the IIFE (uses the concept of closure so it can
still be accessed when the placeMark function is called but cannot be
manipulated otherwise). I also changed the factory function to a
constructor for the players because I thought this would be more simple
for this type of game.