Create check winner/update announcement functions
I created the functions checkOWinner and checkXWinner to go through
all possible variations of winning and checking to see if any are
fulfilled, otherwise the function returns false and the gameOver
variable stays false to allow the next player to continue. If one of
these functions returns true, then the winner is displayed and the
gameOver variable becomes true preventing a next move from being made
until the board is reset. I also created an updateAnnouncement function
which updates the announcement near the top of the screen after every
move to display whose turn it is. This function will also update to show
the winner if there is a winner. The logic for a draw is not yet
completed.