Add functionality to check for a draw
Added two new functions, isTilefull and isBoardFull, whereby the former
function checks if a tile has a mark and the latter checks if the entire
board has a mark. If this is the case and there is no winning play, the
gameOver variable is set to true and the "It's a draw!" message is
displayed signaling a game over.